\name{mc.sim}
\alias{mc.sim}

\title{
Generating next state
}
\description{
Function for generating next state after x1 from a transition matrix P. 
}
\usage{
mc.sim(P,x1)
}

\arguments{
  \item{P}{Transition matrix.
}
\item{x1}{
Numeric value that represents the current state.
}
}

\value{
A numeric value that represents the next state after x1.
}
